pyproject.toml
[tool.poetry]
name = "qqbot"
version = "0.1.0"
description = ""
authors = ["catusax <root@coolrc.me>"]
license = "GPLV3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot-adapter-onebot = "^2.4.3"
nonebot2 = {extras = ["fastapi", "httpx", "websockets"], version = "^2.4.1"}
nonebot-plugin-status = "^0.8.1"
nonebot-plugin-plus-one = "^0.1.10"
numpy = "^1.23.5"
nonebot-plugin-memes = "^0.6.2"
httpx = {extras = ["socks"], version = "^0.28.1"}
openai = "^1.59.3"
nonebot-plugin-alconna = "^0.54.2"
pyyaml = "^6.0.2"
asyncpg = "^0.30.0"


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

